% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greetings.R
\name{cpo_hello}
\alias{cpo_hello}
\title{Say hello to CPO cars}
\usage{
cpo_hello(x)
}
\arguments{
\item{x}{This can be anything. We don't care what you have to say. It's meaningless.}
}
\value{
You give anything to the function (including nothing) Then, you receive a "Hello" from the function.
}
\description{
This \code{cpo_hello()} help us say hello to all the CPO cars!
There were 6 CPO (Certified Pre-owned) cars in a lake,
they did not realize the surrounding water until someone
revealed the fact to them. May be to late to fix them,
there is still time to say hello.
}
\examples{
1+1
cpo_hello(10)
}
