% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_operation.R
\name{full_replace}
\alias{full_replace}
\title{Replace all x in a tibble with y}
\usage{
full_replace(table, target, replacement)
}
\arguments{
\item{table}{This needs to be a  \code{dplyr}  \code{tibble}.}

\item{target}{The target value to be replaced. This can be number, string, or NA.}

\item{replacement}{The value that will be used to replace the \code{target}.}
}
\value{
This function will return a \code{tibble} with all the \code{x} replaced by \code{y}.
}
\description{
While the \code{dplyr} grammar provides great functionality to
replace any cell value, \code{full_replace} makes code more readable and intuitive.
}
\examples{
data = tibble(a=1:10, b=11:20, c='Test')
full_replace(table = data,target='Test',replacement=100)

data = tibble(a=1:10, b=11:20, c=NA)
full_replace(table = data,target=NA,replacement=100)
}
