% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_operation.R
\name{rownonmiss}
\alias{rownonmiss}
\title{Counting the number of non-missing cells within a row}
\usage{
rownonmiss(table)
}
\arguments{
\item{table}{Takes a dataframe or tibble.}
}
\value{
This function will return a vector with the count of all the non-missing cells in each row.
}
\description{
Inspired by \code{STATA}'s \code{rownonmiss} function.
}
\examples{
library(tibble)
library(dplyr)
data = tibble(a=c(NA,NA,1:3), b=rep(NA,5), c=c(100:97,NA))
rownonmiss(data)

}
